+++ part1basic/Expressions.scala
@@ -1,4 +1,4 @@
-package lectures
+package lectures.part1basic
 
 object Expressions extends App{
